Bug correction and attempted fix for log file when running unit tests.
