+++ Ironstone Core Tests/Civil3DAttributeTests.cs
@@ -1,20 +1,21 @@
 ﻿using System.IO;
 using System.Reflection;
+using Autodesk.AutoCAD.ApplicationServices.Core;
 using Autodesk.AutoCAD.Runtime;
+using Jpp.Ironstone.Core.Properties;
 using NUnit.Framework;
 
 namespace Jpp.Ironstone.Core.Tests
 {
     [TestFixture]
-    // TODO: Add a test for whne civil 3d is running to check this works ok. Need Civil 3d test environment to implement 
-    class IronstoneCommandAttributeTests : IronstoneTestFixture
+    class Civil3DCommandAttributeTests : IronstoneTestFixture
     {
-        public IronstoneCommandAttributeTests() : base(Assembly.GetExecutingAssembly(), typeof(IronstoneCommandAttributeTests)) { }
+        public Civil3DCommandAttributeTests() : base(Assembly.GetExecutingAssembly(), typeof(Civil3DCommandAttributeTests)) { }
 
         [Test]
-        public void LoggerCalled()
+        public void Civil3DFailed()
         {
-            bool result = RunTest<bool>(nameof(LoggerCalledResident));
+            bool result = RunTest<bool>(nameof(Civil3DFailedResident));
 
             Configuration config = new Configuration();
             config.TestSettings();
@@ -25,12 +26,14 @@ namespace Jpp.Ironstone.Core.Tests
                 contents = tr.ReadToEnd();
             }
 
-            StringAssert.Contains("LoggerCalledResident", contents);
+            // TODO: Review referencing the resource string directly
+            StringAssert.Contains("Command is not available unless running in Civil 3D", contents);
         }
 
         [IronstoneCommand]
-        [CommandMethod("LoggerCalledResident")]
-        public bool LoggerCalledResident()
+        [Civil3D]
+        [CommandMethod("Civil3DFailedResident")]
+        public bool Civil3DFailedResident()
         {
             return true;
         }
