+++ Ironstone Core Tests/Ironstone Core Tests.csproj
@@ -37,15 +37,15 @@
   <ItemGroup>
     <Reference Include="accoremgd, Version=23.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Jpp.Ironstone.BaseApi.2019.0.0.7\lib\net47\accoremgd.dll</HintPath>
-      <Private>False</Private>
+      <Private>True</Private>
     </Reference>
     <Reference Include="Acdbmgd, Version=23.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Jpp.Ironstone.BaseApi.2019.0.0.7\lib\net47\Acdbmgd.dll</HintPath>
-      <Private>False</Private>
+      <Private>True</Private>
     </Reference>
     <Reference Include="acdbmgdbrep, Version=23.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Jpp.Ironstone.BaseApi.2019.0.0.7\lib\net47\acdbmgdbrep.dll</HintPath>
-      <Private>False</Private>
+      <Private>True</Private>
     </Reference>
     <Reference Include="AcTestFramework, Version=1.0.0.0, Culture=neutral, processorArchitecture=AMD64">
       <HintPath>..\packages\AcTestFramework.2.1.43\lib\net472\AcTestFramework.dll</HintPath>
@@ -101,8 +101,8 @@
     <Compile Include="Autocad\DocumentStoreTests.cs" />
     <Compile Include="Autocad\DrawingObjectManagerTests.cs" />
     <Compile Include="Autocad\DrawingObjectTests.cs" />
-    <Compile Include="Civil3DAttributeTests.cs" />
     <Compile Include="IronstoneCommandAttributeTests.cs" />
+    <Compile Include="Civil3DAttributeTests.cs" />
     <Compile Include="IronstoneTestFixture.cs" />
     <Compile Include="CoreExtensionApplicationTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
