+++ Ironstone Core/ServiceInterfaces/Loggers/FileLogger.cs
@@ -12,7 +12,7 @@ namespace Jpp.Ironstone.Core.ServiceInterfaces.Loggers
 
         public FileLogger(string path)
         {
-            var config = new NLog.Config.LoggingConfiguration();
+            var config = LogManager.Configuration ?? new NLog.Config.LoggingConfiguration();
             var logfile = new NLog.Targets.FileTarget
             {
                 Name= "logfile",
@@ -22,11 +22,13 @@ namespace Jpp.Ironstone.Core.ServiceInterfaces.Loggers
                 MaxArchiveFiles = 10
             };
 
-            config.AddRule(LogLevel.Trace, LogLevel.Fatal, logfile);
+            config.AddTarget(logfile);
+
+            config.AddRule(LogLevel.Trace, LogLevel.Fatal, "logfile", "Ironstone");
 
             LogManager.Configuration = config;
 
-            _logger = LogManager.GetCurrentClassLogger();
+            _logger = LogManager.GetLogger("Ironstone");
         }
 
         public void Dispose()
@@ -63,12 +65,11 @@ namespace Jpp.Ironstone.Core.ServiceInterfaces.Loggers
 
         public override void LogEvent(Event eventType, string eventParameters)
         {
-            _logger.Trace($"{Process.Id}:{eventType}:{eventParameters}");
+            _logger.Error($"{Process.Id}:{eventType}:{eventParameters}");
         }
 
         public override void LogException(Exception exception)
         {
-            //_logger.Error(exception, Process.Id.ToString);
             _logger.Error(exception);
         }
